+++ lib/views/manage_users/manage_users.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:thc/views/widgets.dart';
 
-class UserManagement extends StatelessWidget {
-  const UserManagement({super.key});
+class ManageUsers extends StatelessWidget {
+  const ManageUsers({super.key});
 
   /// The following two functions are equivalent:
   /// ```dart
@@ -14,5 +14,6 @@ class UserManagement extends StatelessWidget {
   /// for methods like [build], since no function body means
   /// you can't just quickly add stuff in when you need to.
   @override
-  Widget build(BuildContext context) => const FunPlaceholder('User Management (add/remove users & assign permissions)!');
+  Widget build(BuildContext context) =>
+      const FunPlaceholder('User Management (add/remove users & assign permissions)!');
 }
