+++ lib/views/survey/fun_quiz.dart
@@ -11,7 +11,7 @@ library;
 import 'dart:math';
 
 import 'package:flutter/material.dart';
-import 'package:thc/models/navigation.dart';
+import 'package:thc/models/navigator.dart';
 import 'package:thc/views/survey/survey_questions.dart';
 import 'package:thc/views/widgets.dart';
