+++ lib/views/survey/survey_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:thc/models/bloc.dart';
-import 'package:thc/models/navigation.dart';
+import 'package:thc/models/navigator.dart';
 import 'package:thc/models/theme.dart';
 import 'package:thc/views/survey/fun_quiz.dart';
 import 'package:thc/views/survey/survey_field.dart';
